Move command RPC utility functions into cliutil
