+++ pkg/client/cli/cliutil/flagtypes_test.go
@@ -1,4 +1,4 @@
-package extensions
+package cliutil
 
 import (
 	"encoding/json"
