+++ pkg/client/cli/command_group.go
@@ -7,30 +7,17 @@ import (
 	"github.com/moby/term"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
+	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/cliutil"
 )
 
-// CommandGroup represents a group of commands and the name of that group
-type CommandGroup struct {
-	Name     string
-	Commands []*cobra.Command
-}
-
-type CommandGroups map[string][]*cobra.Command
-
-// FlagGroup represents a group of flags and the name of that group
-type FlagGroup struct {
-	Name  string
-	Flags *pflag.FlagSet
-}
-
-var commandGroupMap = make(map[string]CommandGroups)
-var globalFlagGroups []FlagGroup
+var commandGroupMap = make(map[string]cliutil.CommandGroups)
+var globalFlagGroups []cliutil.FlagGroup
 
 func init() {
-	cobra.AddTemplateFunc("commandGroups", func(cmd *cobra.Command) CommandGroups {
+	cobra.AddTemplateFunc("commandGroups", func(cmd *cobra.Command) cliutil.CommandGroups {
 		return commandGroupMap[cmd.Name()]
 	})
-	cobra.AddTemplateFunc("globalFlagGroups", func() []FlagGroup {
+	cobra.AddTemplateFunc("globalFlagGroups", func() []cliutil.FlagGroup {
 		return globalFlagGroups
 	})
 	cobra.AddTemplateFunc("wrappedFlagUsages", func(flags *pflag.FlagSet) string {
@@ -66,13 +53,13 @@ func init() {
 	})
 }
 
-func setCommandGroups(cmd *cobra.Command, groups CommandGroups) {
+func setCommandGroups(cmd *cobra.Command, groups cliutil.CommandGroups) {
 	commandGroupMap[cmd.Name()] = groups
 }
 
-// AddCommandGroups adds all the groups in the given CommandGroup to the command,  replaces
+// AddCommandGroups adds all the groups in the given CommandGroups to the command,  replaces
 // the its standard usage template with a template that groups the commands according to that group.
-func AddCommandGroups(cmd *cobra.Command, groups CommandGroups) {
+func AddCommandGroups(cmd *cobra.Command, groups cliutil.CommandGroups) {
 	for _, group := range groups {
 		cmd.AddCommand(group...)
 	}
